move context to fuse
