+++ fuse/__init__.py
@@ -10,3 +10,4 @@ from .vertical_merger_plugin import *
 from .volume_plugin import *
 
 from .context import *
+from .context_utils import *
